Upgraded OVR version to Oculus Mobile SDK 28.0 (1.45.0). Removed option for the Vulkan API dump layer which was never used, and it was officially removed from the example apps in the original SDK. Enabled Phase Sync in the OVR version, to allow for improvements in performance.

+++ SlipNFrag-OVR/app/src/main/cpp/AppState.h
@@ -26,7 +26,7 @@ struct AppState
 	int DefaultFOV;
 	int FOV;
 	Scene Scene;
-	long long FrameIndex;
+	uint64_t FrameIndex;
 	double DisplayTime;
 	int SwapInterval;
 	int CpuLevel;
